Add config to do webpack
Note: I was also making changes to the old main.ts when I started doing this and it is currently broken.

I was attempting to use libs to remove the parts of the SVG paths that were hidden by the clip path as that isn't supported is my preferred gcode generator.